% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIIV.R
\name{plotIIV}
\alias{plotIIV}
\title{Plot intraindividual variability of reaction time}
\usage{
plotIIV(my_data, data_type, block_name, trial_blocks, session_id, trial_number,
  trial_latency)
}
\arguments{
\item{my_data}{The raw dataframe to be used}

\item{data_type}{A string of "raw" for no cleaning, or "clean" for cleaned data (no error trials, RT < 10,000ms, and RT > 180ms)}

\item{block_name}{A string of the variable name for the blocks}

\item{trial_blocks}{A vector of the four essential blocks in the seven-block IAT (i.e., B3, B4, B6, and B7).}

\item{session_id}{A string of the variable name identifying each unique participant.}

\item{trial_number}{A string of the variable identifying the trial number.}

\item{trial_latency}{A string of the variable name for the latency of each trial.}
}
\description{
Plot intraindividual variability in reaction time, faceted by the four essential blocks.
}

