% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotItemErr.R
\name{plotItemErr}
\alias{plotItemErr}
\title{Plot proportion of errors per item in the IAT}
\usage{
plotItemErr(my_data, item_name, trial_latency, trial_error)
}
\arguments{
\item{my_data}{The raw dataframe to be used}

\item{item_name}{A string of the variable identifying the items}

\item{trial_latency}{A string of the variable name for the latency of each trial.}

\item{trial_error}{A string of the variable name identifying whether a trial was an error or not (1 = error)}
}
\description{
Plot proportion of errors in the IAT to see if any items yield high error rates. The data is automatically cleaned
(i.e., trials with RT > 10000 or < 180 are deleted) to avoid over/underinflation of mean error estimates
}

