% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotIndVar}
\alias{plotIndVar}
\title{Plot individual variability in the IAT}
\usage{
plotIndVar(myData, blockName, trialBlocks, sessionID, trialLatency, trialError)
}
\arguments{
\item{myData}{The raw dataframe to be used}

\item{blockName}{A string of the variable name for the blocks}

\item{trialBlocks}{A vector of the four essential blocks in the seven-block IAT (i.e., B3, B4, B6, and B7).}

\item{sessionID}{A string of the variable name identifying each unique participant.}

\item{trialLatency}{A string of the variable name for the latency of each trial.}

\item{trialError}{A string of the variable name identifying whether a trial was an error or not (1 = error)}
}
\description{
This function uses ggplot2 to plot mean participant reaction time with 95\% confidence intervals to see how reaction time varies by participant. The data is automatically cleaned
(i.e., no error trials, trials with RT > 10000 or < 180 are deleted) to avoid over/underinflation of mean estimates and only include trials from
essential blocks.
}

